[BSD] Adds new branch "openbsd".
There was just too many things that would have to be checked, so I
split the files into separate branches for separate operating systems
where it is needed.

Features will be added depending on what system I am working on, and
operating independent features will be merged across the branches as
needed.